# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/me": {
      "get": {
        "operationId": "IdentityController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["identity"]
      }
    },
    "/api/v1/me/profile": {
      "patch": {
        "operationId": "IdentityController_updateProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["identity"]
      }
    },
    "/api/v1/me/skill-passport": {
      "get": {
        "operationId": "IdentityController_skillPassport",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["identity"]
      }
    },
    "/api/v1/me/sessions": {
      "get": {
        "operationId": "IdentityController_sessions",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["identity"]
      }
    },
    "/api/v1/me/sessions/{sessionId}": {
      "delete": {
        "operationId": "IdentityController_revokeSession",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["identity"]
      }
    },
...
    "/api/v1/route-coverages": {
      "get": {
        "operationId": "UnifiedPublicController_routeCoverages",
        "parameters": [
          {
            "name": "country",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["unified-mobility-public"]
      }
    },
    "/api/v1/official-actions": {
      "get": {
        "operationId": "UnifiedPublicController_officialActions",
        "parameters": [
          {
            "name": "country",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["unified-mobility-public"]
      }
    },
...
    "/api/v1/me/official-actions": {
      "post": {
        "operationId": "UnifiedApplicantController_officialAction",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["unified-mobility-applicant"]
      }
    },
    "/api/v1/me/application-qa": {
      "post": {
        "operationId": "UnifiedApplicantController_applicationQa",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["unified-mobility-applicant"]
      }
    },
    "/api/v1/me/mobility-roi": {
      "post": {
        "operationId": "UnifiedApplicantController_mobilityRoi",
        "parameters": [
          {
            "name": "caseId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["unified-mobility-applicant"]
      }
    },
    "/api/v1/me/deadlines": {
      "get": {
        "operationId": "UnifiedApplicantController_deadlines",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["unified-mobility-applicant"]
      },
      "post": {
        "operationId": "UnifiedApplicantController_createDeadline",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["unified-mobility-applicant"]
      }
    },
    "/api/v1/me/saved-items": {
      "get": {
        "operationId": "UnifiedApplicantController_savedItems",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["unified-mobility-applicant"]
      },
      "post": {
        "operationId": "UnifiedApplicantController_saveItem",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["unified-mobility-applicant"]
      }
    },
    "/api/v1/me/copilot": {
      "post": {
        "operationId": "UnifiedApplicantController_copilot",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["unified-mobility-applicant"]
      }
    },
    "/api/v1/me/journey-cases/{id}/command-center": {
      "get": {
        "operationId": "UnifiedApplicantController_commandCenter",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["unified-mobility-applicant"]
      }
    },
    "/api/v1/me/freshness-dashboard": {
      "get": {
        "operationId": "UnifiedApplicantController_freshness",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["unified-mobility-applicant"]
      }
    },
...
  },
  "info": {
    "title": "ProbashJatra API",
    "description": "Safe migration and global opportunity platform API.\n\nGuarantees this API is built to keep:\n- eligibility answers are deterministic and source-backed, and \"unknown\" is a real answer;\n- a verification level is derived from checked facts, never asserted;\n- money moves only on provider-confirmed webhooks, and settles only against verified milestones;\n- synthetic development records are labelled and are not served outside development.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}